+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Unreleased Changes
 ------------------
 
+* Feature - [Major Version] Bump major version to 1.0.0.
+
+* Issue - Drop support for any non-aws prefixed configuration keys.
+
+* Issue - Drop support for the `:dynamodb_store` configuration name.
+
+* Feature - Bump `aws-sessionstore-dynamodb` dependency to `~> 3`.
+
 0.1.0 (2024-11-16)
 ------------------